Remove getResult and move all pixeppipline interfaces to asyncabc
